Rename the amazon_msk directory to amazon_kafka, and apply necessary changes.
