+++ amazon_kafka/assets/configuration/spec.yaml
