+++ amazon_kafka/assets/dashboards/overview.json
