+++ amazon_kafka/assets/service_checks.json
