+++ amazon_kafka/datadog_checks/amazon_kafka/data/conf.yaml.example
