+++ amazon_kafka/manifest.json
