+++ amazon_kafka/tests/fixtures/list_nodes.json
