+++ amazon_kafka/tests/mock_boto3/boto3/__init__.py
