+++ amazon_kafka/tests/mock_boto3/boto3/list_nodes.json
