+++ amazon_kafka/tests/mock_boto3/setup.py
